提取transfo_xl常用方法至utils，并更新相关代码
