+++ fengshen/models/transfo_xl_reasoning/__init__.py
@@ -1,3 +1,3 @@
+# encoding=utf-8
 from fengshen.models.transfo_xl_denoise.modeling_transfo_xl_denoise import TransfoXLDenoiseModel as TransfoXLModel
 from .generate import deduction_generate, abduction_generate
-# from .modeling_transfo_xl import TransfoXLModel
