+++ fengshen/utils/__init__.py
@@ -1,3 +1,4 @@
 from .universal_checkpoint import UniversalCheckpoint
 from .utils import chinese_char_tokenize
-__all__ = ['UniversalCheckpoint', 'chinese_char_tokenize']
+from .transfo_xl_utils import top_k_logits, sample_sequence_batch, sample_sequence, get_masks_and_position_ids
+__all__ = ['UniversalCheckpoint', 'chinese_char_tokenize', 'top_k_logits', 'sample_sequence_batch', 'sample_sequence', 'get_masks_and_position_ids']
